[CIAS30-3498] Extract OldRecipientTable component
