+++ app/containers/Sessions/components/QuestionData/ThirdPartyQuestion/RecipientsChooser/index.ts
@@ -1,0 +1,2 @@
+export * from './RecipientsChooser';
+export * from './types';
