+++ app/containers/Sessions/components/QuestionData/ThirdPartyQuestion/constants.js
@@ -4,5 +4,3 @@ export const ADD = 'app/EditSessionPage/ThirdParty/ADD';
 export const UPDATE_ANSWER = 'app/EditSessionPage/ThirdParty/UPDATE';
 export const REMOVE = 'app/EditSessionPage/ThirdParty/REMOVE';
 export const REORDER = 'app/EditSessionPage/ThirdParty/REORDER';
-
-export const MANAGE_RECIPIENTS_MODAL_WIDTH = '532px';
