+++ app/containers/Sessions/components/QuestionData/ThirdPartyQuestion/messages.js
@@ -40,36 +40,4 @@ export default defineMessages({
     id: `${scope}.reorderIconAlt`,
     defaultMessage: 'Reorder icon {index}',
   },
-  manageRecipients: {
-    id: `${scope}.manageRecipients`,
-    defaultMessage: 'Manage recipients',
-  },
-  emailRecipients: {
-    id: `${scope}.emailRecipients`,
-    defaultMessage: 'E-mail recipients',
-  },
-  noEmailRecipients: {
-    id: `${scope}.noEmailRecipients`,
-    defaultMessage: 'No e-mail recipients',
-  },
-  faxRecipients: {
-    id: `${scope}.faxRecipients`,
-    defaultMessage: 'Fax recipients',
-  },
-  noFaxRecipients: {
-    id: `${scope}.noFaxRecipients`,
-    defaultMessage: 'No fax recipients',
-  },
-  manageRecipientsModalDefaultTitle: {
-    id: `${scope}.manageRecipientsModalDefaultTitle`,
-    defaultMessage: 'Unlabelled answer',
-  },
-  manageRecipientsModalDescription: {
-    id: `${scope}.manageRecipientsModalDescription`,
-    defaultMessage: 'Determine where you want reports to be sent.',
-  },
-  deleteRecipient: {
-    id: `${scope}.deleteRecipient`,
-    defaultMessage: 'Delete recipient',
-  },
 });
